# Replace Principal Moment loop calculation with vectorized NumPy operations